requirements.txt
asgiref
attrs
Django
django-filter
djangorestframework
drf-spectacular
inflection
jsonschema
pyrsistent
pytz
PyYAML
sqlparse
uritemplate
